test(persistence): split SSH remote PTY binding partition cases into their own file
