+++ src/main/persistence-host-partitioned-sessions.test.ts
@@ -17,7 +17,6 @@ import {
   makeWorkspaceLineage
 } from './persistence-test-harness'
 import { worktreeWorkspaceKey } from '../shared/workspace-scope'
-import { TEST_LEAF_1 } from './persistence-session-fixtures'
 
 // Stub the ~/.ssh/config parser so the SSH-import test drives the real Store with deterministic hosts, not the operator's actual ~/.ssh/config.
 const { loadUserSshConfigMock, sshConfigHostsToTargetsMock } = vi.hoisted(() => ({
@@ -93,35 +92,6 @@ describe('Store host-partitioned workspace sessions', () => {
       }
     }
   }
-
-  const makeBoundHostSession = (ptyId: string | null): WorkspaceSessionState => ({
-    ...getDefaultWorkspaceSession(),
-    activeRepoId: 'repo-1',
-    activeWorktreeId: 'repo-1::/worktree',
-    activeTabId: 'tab-1',
-    tabsByWorktree: {
-      'repo-1::/worktree': [
-        {
-          id: 'tab-1',
-          worktreeId: 'repo-1::/worktree',
-          title: 'Terminal',
-          customTitle: null,
-          color: null,
-          sortOrder: 0,
-          createdAt: 1,
-          ptyId
-        }
-      ]
-    },
-    terminalLayoutsByTabId: {
-      'tab-1': {
-        root: { type: 'leaf', leafId: TEST_LEAF_1 },
-        activeLeafId: TEST_LEAF_1,
-        expandedLeafId: null,
-        ptyIdsByLeafId: ptyId ? { [TEST_LEAF_1]: ptyId } : {}
-      }
-    }
-  })
 
   // Registered on purpose: rows owned by an unregistered repo id are swept as orphans on load.
   const makeRepos = (...repoIds: string[]) => repoIds.map((id) => makeRepo({ id, path: `/${id}` }))
@@ -395,108 +365,6 @@ describe('Store host-partitioned workspace sessions', () => {
     expect(
       store.getWorkspaceSession('runtime:env-b').terminalTopologyRevisionByRepoId?.duplicate
     ).toBe(7)
-  })
-
-  it('persists an SSH PTY binding only in the SSH host partition', async () => {
-    const store = await createStore()
-    store.setWorkspaceSession(makeBoundHostSession(null), 'local')
-    store.setWorkspaceSession(makeBoundHostSession(null), 'ssh:ssh-1')
-
-    store.persistPtyBinding(
-      {
-        worktreeId: 'repo-1::/worktree',
-        tabId: 'tab-1',
-        leafId: TEST_LEAF_1,
-        ptyId: 'ssh:ssh-1@@remote-pty'
-      },
-      'ssh:ssh-1'
-    )
-
-    expect(
-      store.getWorkspaceSession('ssh:ssh-1').tabsByWorktree['repo-1::/worktree'][0]?.ptyId
-    ).toBe('ssh:ssh-1@@remote-pty')
-    expect(
-      store.getWorkspaceSession('local').tabsByWorktree['repo-1::/worktree'][0]?.ptyId
-    ).toBeNull()
-  })
-
-  it('rolls back a failed SSH PTY binding flush in the SSH host partition', async () => {
-    const store = await createStore()
-    store.setWorkspaceSession(makeBoundHostSession(null), 'local')
-    store.setWorkspaceSession(makeBoundHostSession(null), 'ssh:ssh-1')
-    const flush = vi.spyOn(store, 'flushOrThrow').mockImplementationOnce(() => {
-      throw new Error('disk unavailable')
-    })
-
-    expect(() =>
-      store.persistPtyBinding(
-        {
-          worktreeId: 'repo-1::/worktree',
-          tabId: 'tab-1',
-          leafId: TEST_LEAF_1,
-          ptyId: 'ssh:ssh-1@@remote-pty'
-        },
-        'ssh:ssh-1'
-      )
-    ).toThrow('disk unavailable')
-    flush.mockRestore()
-
-    expect(
-      store.getWorkspaceSession('ssh:ssh-1').tabsByWorktree['repo-1::/worktree'][0]?.ptyId
-    ).toBeNull()
-    expect(
-      store.getWorkspaceSession('local').tabsByWorktree['repo-1::/worktree'][0]?.ptyId
-    ).toBeNull()
-  })
-
-  it('clears terminated SSH PTY bindings from the SSH partition and legacy local copy', async () => {
-    const store = await createStore()
-    const ptyId = 'ssh:ssh-1@@remote-pty'
-    store.setWorkspaceSession(makeBoundHostSession(ptyId), 'local')
-    store.setWorkspaceSession(makeBoundHostSession(ptyId), 'ssh:ssh-1')
-    store.upsertSshRemotePtyLease({
-      targetId: 'ssh-1',
-      ptyId: 'remote-pty',
-      worktreeId: 'repo-1::/worktree',
-      tabId: 'tab-1',
-      leafId: TEST_LEAF_1,
-      state: 'attached'
-    })
-
-    store.markSshRemotePtyLease('ssh-1', ptyId, 'terminated')
-
-    for (const hostId of ['local', 'ssh:ssh-1']) {
-      const session = store.getWorkspaceSession(hostId)
-      expect(session.tabsByWorktree['repo-1::/worktree'][0]?.ptyId).toBeNull()
-      expect(session.terminalLayoutsByTabId['tab-1']?.ptyIdsByLeafId).toEqual({})
-    }
-  })
-
-  // `expired` only records that the client lost its route, so both partitions keep the binding the
-  // pane needs to reattach to a remote shell nothing has attested is dead.
-  it('keeps expired SSH PTY bindings in the SSH partition and legacy local copy', async () => {
-    const store = await createStore()
-    const ptyId = 'ssh:ssh-1@@remote-pty'
-    store.setWorkspaceSession(makeBoundHostSession(ptyId), 'local')
-    store.setWorkspaceSession(makeBoundHostSession(ptyId), 'ssh:ssh-1')
-    store.upsertSshRemotePtyLease({
-      targetId: 'ssh-1',
-      ptyId: 'remote-pty',
-      worktreeId: 'repo-1::/worktree',
-      tabId: 'tab-1',
-      leafId: TEST_LEAF_1,
-      state: 'attached'
-    })
-
-    store.markSshRemotePtyLease('ssh-1', ptyId, 'expired')
-
-    for (const hostId of ['local', 'ssh:ssh-1']) {
-      const session = store.getWorkspaceSession(hostId)
-      expect(session.tabsByWorktree['repo-1::/worktree'][0]?.ptyId).toBe(ptyId)
-      expect(session.terminalLayoutsByTabId['tab-1']?.ptyIdsByLeafId).toEqual({
-        [TEST_LEAF_1]: ptyId
-      })
-    }
   })
 
   it('defaults an omitted hostId to the local partition', async () => {
